refactor(ui): move data-table chrome into components/table
Keep ShadCN Table in components/ui; shared list surface, empty shell, and chrome helpers belong with the existing table kit.
