+++ components/table/index.ts
@@ -1,4 +1,14 @@
 export {
+  dataTableBodyCellClassName,
+  dataTableBodyRowClassName,
+  dataTableHeaderCellClassName,
+} from "./data-table-chrome";
+export { DataTableEmpty } from "./data-table-empty";
+export {
+  DATA_TABLE_SURFACE_CLASS_NAME,
+  DataTableSurface,
+} from "./data-table-surface";
+export {
   createPagedTableFooterProps,
   PagedTableFooter,
   type PagedTableFooterProps,
